Modify Module Id Logic in identification when there are multiple org units
Current Behaviour : For identification calls, module id is determined using the org unit hierarchy of the DHIS user who is logged in. If a DHIS user is tagged to multiple org units, the module id is passed as NA.

New Requirement : For identification calls, module id is determined using the org unit hierarchy of the DHIS user who is logged in. If a DHIS user is tagged to the multiple org units, find the unique level 4 and above org units and pass the sub district id if only 1 sub district is retrieved, else pass NA.